Add files via upload
Fitted a Logistic regression model to classify whether a person is diabetic or non-diabetic subject to certain independent variables. Performed all the required data-preprocessing like missing value removal, outlier removal and data visualization. Split the dataset into a train and test set and fit a logistic regression model on the train set. Based on this fitted model, calculated the predicted outcomes of the test set and compared it with that of the actual outcomes of the test set. Enumerated various metrics on this fit ,like accuracy,recall,precision,F1 score, AUC and ROC to make necessary interpretations.